chore: moved unit tests to nl.info package
