--- src/test/kotlin/net/atos/zac/app/identity/model/IdentityFixtures.kt
@@ -1,16 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-
-package net.atos.zac.app.identity.model
-
-import nl.info.zac.app.identity.model.RestUser
-
-fun createRESTUser(
-    id: String = "dummyId",
-    name: String = "dummyUserName"
-) = RestUser(
-    id = id,
-    naam = name
-)
