+++ src/test/kotlin/nl/info/zac/app/model/IdentityFixtures.kt
@@ -1,0 +1,16 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Lifely
+ * SPDX-License-Identifier: EUPL-1.2+
+ */
+
+package nl.info.zac.app.model
+
+import nl.info.zac.app.identity.model.RestUser
+
+fun createRESTUser(
+    id: String = "dummyId",
+    name: String = "dummyUserName"
+) = RestUser(
+    id = id,
+    naam = name
+)
