+++ src/test/kotlin/nl/info/zac/app/zaak/converter/RestDecisionConverterTest.kt
@@ -13,7 +13,6 @@ import io.mockk.every
 import io.mockk.mockk
 import net.atos.client.zgw.drc.DrcClientService
 import net.atos.zac.app.informatieobjecten.converter.RestInformatieobjectConverter
-import net.atos.zac.app.informatieobjecten.model.createRestEnkelvoudigInformatieobject
 import nl.info.client.zgw.brc.BrcClientService
 import nl.info.client.zgw.brc.model.createBesluit
 import nl.info.client.zgw.brc.model.generated.VervalredenEnum
@@ -23,6 +22,7 @@ import nl.info.client.zgw.model.createZaak
 import nl.info.client.zgw.util.extractUuid
 import nl.info.client.zgw.ztc.ZtcClientService
 import nl.info.client.zgw.ztc.model.createBesluitType
+import nl.info.zac.app.informatieobjecten.model.createRestEnkelvoudigInformatieobject
 import nl.info.zac.app.zaak.model.createRestDecisionCreateData
 import nl.info.zac.configuratie.ConfiguratieService
 import java.time.LocalDate
